Fixed a number of compiler warnings.

+++ libosmscout-import/src/osmscout/import/GenOptimizeAreaWayIds.cpp
@@ -279,7 +279,7 @@ namespace osmscout {
             nodeIds.insert(node.GetId());
           }
 
-          idCount+=nodeIds.size();
+          idCount+=static_cast<uint32_t>(nodeIds.size());
 
           for (auto nodeId : nodeIds) {
             if (usedIdSet.find(nodeId)!=usedIdSet.end()) {
@@ -346,7 +346,7 @@ namespace osmscout {
           nodeIds.insert(node.GetId());
         }
 
-        idCount+=nodeIds.size();
+        idCount+=static_cast<uint32_t>(nodeIds.size());
 
         for (auto id : nodeIds) {
           if (usedIdSet.find(id)!=usedIdSet.end()) {
